refactor: move pydantic models to a separate schemas file
