+++ backend/app/auth/schemas.py
@@ -1,0 +1,23 @@
+from pydantic import BaseModel, ConfigDict, EmailStr
+
+
+class Token(BaseModel):
+    access_token: str
+    token_type: str
+
+
+class TokenData(BaseModel):
+    email: EmailStr
+
+
+class UserBase(BaseModel):
+    email: EmailStr
+
+
+class UserCreate(UserBase):
+    password: str
+
+
+class UserPublic(UserBase):
+    model_config = ConfigDict(from_attributes=True)
+    id: int
